[ALPS02124274] pinctrl: mediatek: fix build warning
[Detail] pinctrl: mediatek: fix build warning
/worktmp/src_alps-trunk-m0.basic_kernel/kernel-3.18/
drivers/pinctrl/mediatek/pinctrl-mtk-common.c:1424:8:
note: in expansion of macro DEVICE_ATTR
static DEVICE_ATTR(mt_gpio, 0664, mt_gpio_show_pin, mt_gpio_store_pin);
        ^

[Solution] modify return type of function
mt_gpio_show and mt_gpio_store in pinctrl-mtk-common.c

[Feature] Others

MTK-Commit-Id: cf1e0df86735e4c14d0011bb295b7edd2c6499c8

CR-Id: ALPS02124274

+++ drivers/pinctrl/mediatek/pinctrl-mtk-common.c
@@ -1363,7 +1363,7 @@ static int mtk_driving_get(struct gpio_chip *chip, unsigned offset)
 	return ((val & mask) >> shift);
 }
 
-static int mt_gpio_show_pin(struct device *dev, struct device_attribute *attr, char *buf)
+static ssize_t mt_gpio_show_pin(struct device *dev, struct device_attribute *attr, char *buf)
 {
 	int len = 0;
 	int bufLen = PAGE_SIZE;
@@ -1396,7 +1396,7 @@ static int mt_gpio_show_pin(struct device *dev, struct device_attribute *attr, c
 	return len;
 }
 
-static int mt_gpio_store_pin(struct device *dev, struct device_attribute *attr, const char *buf, size_t count)
+static ssize_t mt_gpio_store_pin(struct device *dev, struct device_attribute *attr, const char *buf, size_t count)
 {
 	int pin, val;
 	int val_set;
